+++ mundo-web/src/test/java/cn/fantasticmao/mundo/web/support/JsonApiTest.java
@@ -1,7 +1,11 @@
 package cn.fantasticmao.mundo.web.support;
 
+import cn.fantasticmao.mundo.core.util.JsonUtil;
+import com.fasterxml.jackson.core.JacksonException;
+import com.fasterxml.jackson.core.type.TypeReference;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
+import org.springframework.http.HttpStatus;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -16,53 +20,65 @@ import java.util.Map;
 public class JsonApiTest {
 
     @Test
-    public void testJsonApi1() {
+    public void fromJson() throws JacksonException {
+        String json = "{\"status\":false,\"code\":404,\"message\":\"Not Found\",\"data\":\"hello\"}";
+        TypeReference<JsonApi<String>> reference = new TypeReference<>() {
+        };
+        JsonApi<String> jsonApi = JsonUtil.fromJson(json, reference);
+        Assertions.assertFalse(jsonApi.isStatus());
+        Assertions.assertEquals(HttpStatus.NOT_FOUND.value(), jsonApi.getCode());
+        Assertions.assertEquals(HttpStatus.NOT_FOUND.getReasonPhrase(), jsonApi.getMessage());
+        Assertions.assertEquals("hello", jsonApi.getData());
+    }
+
+    @Test
+    public void toJson1() throws JacksonException {
         String expected = "{\"status\":true,\"code\":200,\"message\":\"OK\"}";
-        String actual = JsonApi.success().data(null).toString();
+        String actual = JsonUtil.toJson(JsonApi.success().data(null));
         Assertions.assertEquals(expected, actual);
     }
 
     @Test
-    public void testJsonApi2() {
+    public void toJson2() throws JacksonException {
         String expected = "{\"status\":true,\"code\":200,\"message\":\"OK\",\"data\":1}";
-        String actual = JsonApi.success().data(1).toString();
+        String actual = JsonUtil.toJson(JsonApi.success().data(1));
         Assertions.assertEquals(expected, actual);
     }
 
     @Test
-    public void testJsonApi3() {
+    public void toJson3() throws JacksonException {
         String expected = "{\"status\":true,\"code\":200,\"message\":\"OK\",\"data\":true}";
-        String actual = JsonApi.success().data(true).toString();
+        String actual = JsonUtil.toJson(JsonApi.success().data(true));
         Assertions.assertEquals(expected, actual);
     }
 
     @Test
-    public void testJsonApi4() {
+    public void toJson4() throws JacksonException {
         String expected = "{\"status\":true,\"code\":200,\"message\":\"OK\",\"data\":\"hello\"}";
-        String actual = JsonApi.success().data("hello").toString();
+        String actual = JsonUtil.toJson(JsonApi.success().data("hello"));
         Assertions.assertEquals(expected, actual);
     }
 
     @Test
-    public void testJsonApi5() {
+    public void toJson5() throws JacksonException {
         String expected = "{\"status\":true,\"code\":200,\"message\":\"OK\",\"data\":[1,2,3]}";
-        String actual = JsonApi.success().data(new int[]{1, 2, 3}).toString();
+        String actual = JsonUtil.toJson(JsonApi.success().data(new int[]{1, 2, 3}));
         Assertions.assertEquals(expected, actual);
     }
 
     @Test
-    public void testJsonApi6() {
+    public void toJson6() throws JacksonException {
         String expected = "{\"status\":true,\"code\":200,\"message\":\"OK\",\"data\":{\"hello\":123}}";
         Map<String, Integer> map = new HashMap<>();
         map.put("hello", 123);
-        String actual = JsonApi.success().data(map).toString();
+        String actual = JsonUtil.toJson(JsonApi.success().data(map));
         Assertions.assertEquals(expected, actual);
     }
 
     @Test
-    public void testJsonApi7() {
+    public void toJson7() throws JacksonException {
         String expected = "{\"status\":true,\"code\":200,\"message\":\"OK\",\"data\":false}";
-        String actual = JsonApi.success().data(true).data(false).toString();
+        String actual = JsonUtil.toJson(JsonApi.success().data(true).data(false));
         Assertions.assertEquals(expected, actual);
     }
 }